Implement search functionality with query param
