+++ app/todo/[id]/page.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { client } from '@/sanity/lib/client';
 import PageNotFound from '@/components/UI/PageNotFound';
 import Link from 'next/link';
-import { FaFacebook, FaInstagram, FaPinterest } from 'react-icons/fa';
 
 type PropTypes = {
   params: { id: string };
